happy/requirements.txt
asn1crypto==0.24.0
boto3==1.9.4
botocore==1.12.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-appconf==1.0.2
django-avatar==4.1.0
django-cors-headers==2.3.0
django-extensions==2.0.7
django-filter==2.0.0
django-rest-auth==0.9.3
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.90.2
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
oauthlib==2.1.0
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
whitenoise==3.3.1
